[dev8] Factor out handleSimulatorEvents
